update style -> make preview container scrollable, change scroll bar style. can consider this plugin done.
Future work/improvement: currently the list of images are collected by
providing a jQuery selector to these imagee. In the next version,
developers can
put a "data-image-gallery-plugin" attribute to every image element and
this plugin can do its work. This would be better because in some
situation, it's complicated to provide a selector for all the image
element.

Further future work: can re-write the plugin to make it jQuery-free, so
that it can be used where a project does not rely on jQuery.